Replaced System.Drawing with SixLabors' ImageSharp library. This produces better results, especially when it comes to darker gradients (likely gamma-related).
ImageSharp comes with its own quantization/dithering 'pipeline', but that approach makes it difficult to obtain the final texture data,
and would require custom implementations anyway due to transparent and water texture handling.
The only idea taken from ImageSharp's dithering is to use an error scale instead of a threshold.