Created a GetTextureName function.
The idea was to ignore certain parts of a filename, which would make wildcard rules more powerful (rename a file to switch to different settings, without changing the output texture name),
but renaming a file won't update its last-modified-time, so that would require a more elaborate mechanism for keeping track of previous settings, file contents, etc.
Note that WadMaker currently isn't able to detect 'filename swapping' (rename A to B, and B to A, without otherwise modifying them), so it's already a bit of an issue.