Added support for making fullbright textures: if a texture name starts with a ~ (or +N~, for animated textures), then the last 32 colors are used for fullbright pixels.
To set fullbright pixels, create a "{texturename}.fullbright" mask image that is fully transparent, except for the pixels that should be fullbright.
The alpha threshold for this fullbright mask image can be adjusted with a "fullbright-alpha-threshold: N" rule in a wadmaker.config file.
Fullbright support can be disabled with a "no-fullbright: true" rule in a wadmaker.config file.

This feature is intended for Quake engines that support WAD3 wad files, because fullbright textures are not supported in GoldSource.